requirements.txt
lxml==4.2.5
matplotlib==2.2.2
numpy==1.15.4
opencv-python==3.4.0.12
pandas==0.23.4
Pillow==5.1.0
protobuf==3.6.1
scikit-image==0.14.2
scikit-learn==0.19.2
scipy==1.0.1
six==1.12.0
sklearn==0.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
tensorflow-tensorboard==0.4.0